chore(test): rename test/testing.py to test/common.py and exclude tests from wheel
- Rename test/testing.py to test/common.py per review feedback
- Update all imports across trainer, spark, optimizer, hub, and common tests
- Update AGENTS.md/CLAUDE.md references
- Exclude *_test.py files from the built wheel via hatch build config
