+++ kubeflow/optimizer/backends/kubernetes/backend_test.py
@@ -57,7 +57,7 @@ from kubeflow.trainer.types.types import (
     TrainJob,
     TrainJobTemplate,
 )
-from test.testing import (
+from test.common import (
     DEFAULT_NAMESPACE,
     FAILED,
     RUNTIME,
