+++ kubeflow/spark/backends/kubernetes/utils_test.py
@@ -53,7 +53,7 @@ from kubeflow.spark.types.types import (
     SparkConnectState,
     SparkJobStatus,
 )
-from test.testing import FAILED, SUCCESS, TestCase
+from test.common import FAILED, SUCCESS, TestCase
 
 # --------------------------
 # Fixtures
