+++ kubeflow/trainer/backends/container/runtime_loader_test.py
@@ -25,7 +25,7 @@ import pytest
 from kubeflow.trainer.backends.container import runtime_loader
 from kubeflow.trainer.constants import constants
 from kubeflow.trainer.types import types as base_types
-from test.testing import FAILED, SUCCESS, TestCase
+from test.common import FAILED, SUCCESS, TestCase
 
 # Sample runtime YAML data for testing
 SAMPLE_RUNTIME_YAML = {
