+++ kubeflow/trainer/backends/container/utils_test.py
@@ -18,7 +18,7 @@ from kubeflow.common.constants import UNKNOWN
 from kubeflow.trainer.backends.container import utils as container_utils
 from kubeflow.trainer.constants import constants
 from kubeflow.trainer.types import types
-from test.testing import FAILED, SUCCESS, TestCase
+from test.common import FAILED, SUCCESS, TestCase
 
 
 def simple_train_func():
