+++ kubeflow/trainer/backends/kubernetes/utils_test.py
@@ -23,7 +23,7 @@ import pytest
 import kubeflow.trainer.backends.kubernetes.utils as utils
 from kubeflow.trainer.constants import constants
 from kubeflow.trainer.types import types
-from test.testing import FAILED, SUCCESS, TestCase
+from test.common import FAILED, SUCCESS, TestCase
 
 # --------------------------
 # Test Helpers
